+++ R/app_server.R
@@ -9,7 +9,8 @@ app_server <- function(input, output, session) {
   utils_authorize_GS()
 
   .df <- reactive(fct_load_shovel_df())
-  .col_defs <- reactive(utils_html_col_defs(.df(), index_grey_cols = c(1:2)))
 
-  mod_regen_data_server("regen_data_1", .df, .col_defs)
+  .list_defs <- reactive(utils_list_defs(.df))
+
+  mod_regen_data_server("regen_data_1", .df, .list_defs)
 }