+++ NEWS.md
@@ -1,5 +1,5 @@
 # shovelApp 0.1.0
 
 * Created the package
-* Added `utils`, `fct_`, and `mod` prefixed functions as per the golem package 
+* Added `utils` and `mod` prefixed functions as per the golem package 
 * Added tests